--- sale_import_amazon/models/__init__.py
@@ -1,7 +1,0 @@
-from . import amazon_marketplace
-from . import queue_job_chunk
-from . import sale_channel
-from . import sale_channel_importer_amazon
-from . import sale_order
-
-from . import schemas
