--- sale_import_amazon/security/ir.model.access.csv
@@ -1,3 +1,0 @@
-id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
-sale_import_amazon.access_sale_channel_importer_amazon,access_sale_channel_importer_amazon,sale_import_amazon.model_sale_channel_importer_amazon,base.group_user,1,1,1,1
-sale_import_amazon.access_amazon_marketplace,access_amazon_marketplace,sale_import_amazon.model_amazon_marketplace,base.group_user,1,0,0,0
