+++ sale_import_amazon_base/models/__init__.py
@@ -1,0 +1,4 @@
+from . import amazon_marketplace
+from . import sale_channel
+from . import sale_order
+from . import schemas
