+++ sale_import_amazon_base/models/sale_order.py
@@ -1,0 +1,10 @@
+# Copyright 2024 Akretion
+# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
+
+from odoo import fields, models
+
+
+class SaleOrder(models.Model):
+    _inherit = "sale.order"
+
+    amazon_marketplace_id = fields.Many2one("amazon.marketplace")
