+++ sale_import_amazon_base/security/ir.model.access.csv
@@ -1,0 +1,2 @@
+id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
+sale_import_amazon_base.access_amazon_marketplace,access_amazon_marketplace,sale_import_amazon_base.model_amazon_marketplace,base.group_user,1,0,0,0
