+++ sale_import_amazon_base/views/amazon_marketplace.xml
@@ -1,0 +1,54 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<!-- Copyright 2024 Akretion
+     License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl). -->
+<odoo>
+
+    <record id="amazon_marketplace_form_view" model="ir.ui.view">
+        <field name="name">amazon.marketplace.form</field>
+        <field name="model">amazon.marketplace</field>
+        <field name="arch" type="xml">
+            <form>
+                <header>
+                    <field name="name" />
+                </header>
+                <sheet>
+                    <group>
+                        <group>
+                            <field name="country_code" />
+                            <field name="marketplace_ref" />
+                        </group>
+                    </group>
+                </sheet>
+            </form>
+        </field>
+    </record>
+
+
+    <record id="amazon_marketplace_tree_view" model="ir.ui.view">
+        <field name="name">amazon.marketplace.tree</field>
+        <field name="model">amazon.marketplace</field>
+        <field name="arch" type="xml">
+            <tree>
+                <field name="name" />
+                <field name="country_code" />
+                <field name="marketplace_ref" />
+            </tree>
+        </field>
+    </record>
+
+    <record id="amazon_marketplace_act_window" model="ir.actions.act_window">
+        <field name="name">Amazon Marketplace</field>
+        <field name="res_model">amazon.marketplace</field>
+        <field name="view_mode">tree,form</field>
+        <field name="domain">[]</field>
+        <field name="context">{}</field>
+    </record>
+
+    <record model="ir.ui.menu" id="amazon_marketplace_menu">
+        <field name="name">Amazon Marketplace</field>
+        <field name="parent_id" ref="sale.menu_sale_config" />
+        <field name="action" ref="amazon_marketplace_act_window" />
+        <field name="sequence" eval="99" />
+    </record>
+
+</odoo>
