+++ sale_import_amazon_base/views/sale_channel.xml
@@ -1,0 +1,36 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<!-- Copyright 2024 Akretion
+     License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl). -->
+<odoo>
+
+    <record model="ir.ui.view" id="sale_channel_form_view">
+        <field name="name">sale.channel.form (in sale_import_amazon_base)</field>
+        <field name="model">sale.channel</field>
+        <field name="inherit_id" ref="sale_import_base.sale_channel_view_form" />
+        <field name="arch" type="xml">
+            <xpath expr="//group[@name='sale_import_base']/.." position="after">
+                <group
+                    name="sale_import_amazon"
+                    string="Amazon Import Parameters"
+                    invisible="1"
+                >
+                    <group name="amazon_credentials" groups="base.group_system">
+                        <field name="lwa_appid" />
+                        <field name="sp_api_refresh_token" />
+                        <field name="lwa_client_secret" />
+                    </group>
+                    <group name="amazon_data">
+                        <field
+                            name="marketplace_ids"
+                            widget="many2many_tags"
+                            domain="[('country_code','!=',False)]"
+                        />
+                    </group>
+                </group>
+            </xpath>
+        </field>
+    </record>
+
+
+
+</odoo>
