+++ sale_import_amazon_seller/README.rst
@@ -1,0 +1,23 @@
+============================
+Amazon Connector for Sellers
+============================
+
+Catch Amazon Sale Orders data and created related Odoo Sale Orders
+
+Configuration
+=============
+
+To configure this module, you need to:
+
+1. Create a Sale Channel called "Amazon Seler"
+2. See sale_import_amazon_base for information about credentials
+
+Usage
+=====
+
+To use this module, you need to:
+
+#. Go to ...
+
+
+
