+++ sale_import_amazon_seller/__init__.py
@@ -1,0 +1,1 @@
+from . import models
