+++ sale_import_amazon_seller/models/__init__.py
@@ -1,0 +1,5 @@
+from . import queue_job_chunk
+from . import sale_channel
+from . import sale_channel_importer_amazon
+from . import sale_order
+from . import schemas
