+++ sale_import_amazon_seller/models/queue_job_chunk.py
@@ -1,0 +1,25 @@
+# Copyright 2024 Akretion
+# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
+
+from odoo import fields, models
+
+
+class QueueJobChunk(models.Model):
+    _inherit = "queue.job.chunk"
+
+    processor = fields.Selection(
+        selection_add=[
+            (
+                "sale_channel_importer_amazon_seller",
+                "Sale Channel Importer Amazon Seller",
+            )
+        ]
+    )
+
+    def _get_processor(self):
+        if self.processor == "sale_channel_importer_amazon_seller":
+            return self.env["sale.channel.importer.amazon.seller"].new(
+                {"chunk_id": self.id}
+            )
+        else:
+            return super()._get_processor()
