+++ sale_import_amazon_seller/views/sale_channel.xml
@@ -1,0 +1,28 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<!-- Copyright 2024 Akretion
+     License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl). -->
+<odoo>
+
+    <record model="ir.ui.view" id="sale_channel_form_view">
+        <field name="name">sale.channel.form (in sale_import_amazon_seller)</field>
+        <field name="model">sale.channel</field>
+        <field name="inherit_id" ref="sale_import_amazon_base.sale_channel_form_view" />
+        <field name="arch" type="xml">
+            <xpath expr="//group[@name='sale_import_amazon']" position="inside">
+                <button
+                    name="amazon_seller_import_orders_chunk"
+                    string="Import Amazon Orders"
+                    type="object"
+                    class="btn btn-primary col-6 col-lg-3"
+                />
+            </xpath>
+            <xpath expr="//group[@name='amazon_data']" position="inside">
+                <field name="amazon_location_id" />
+                <field name="date_last_sale_update" />
+            </xpath>
+        </field>
+    </record>
+
+
+
+</odoo>
