+++ sale_import_amazon_seller/views/sale_order.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<!-- Copyright 2024 Akretion
+     License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl). -->
+<odoo>
+
+    <record model="ir.ui.view" id="sale_order_form_view">
+        <field name="name">sale.order.form (in sale_import_amazon_seller)</field>
+        <field name="model">sale.order</field>
+        <field name="inherit_id" ref="sale_import_amazon_base.sale_order_form_view" />
+        <field name="arch" type="xml">
+            <xpath expr="//group[@name='amazon_import']" position="inside">
+                <field name="is_fulfilled_by_amazon" readonly="1" />
+            </xpath>
+        </field>
+    </record>
+
+
+
+</odoo>
