--- setup/sale_import_amazon/odoo/addons/sale_import_amazon
@@ -1,1 +1,0 @@
-../../../../sale_import_amazon
