+++ setup/sale_import_amazon_base/odoo/addons/sale_import_amazon_base
@@ -1,0 +1,1 @@
+../../../../sale_import_amazon_base
