+++ setup/sale_import_amazon_seller/odoo/addons/sale_import_amazon_seller
@@ -1,0 +1,1 @@
+../../../../sale_import_amazon_seller
